+++ collections/posts.js
@@ -507,6 +507,17 @@ Meteor.methods({
         set.postedAt = new Date();
 
       var result = Posts.update(post._id, {$set: set}, {validate: false});
+
+      // --------------------- Server-Side Async Callbacks --------------------- //
+      if (Meteor.isServer) {
+        Meteor.defer(function () { // use defer to avoid holding up client
+          // run all post submit server callbacks on post object successively
+          post = postApproveCallbacks.reduce(function(result, currentFunction) {
+              return currentFunction(result);
+          }, post);
+        });
+      }
+
     }else{
       flashMessage('You need to be an admin to do that.', "error");
     }