+++ packages/telescope-notifications/lib/herald.js
@@ -1,3 +1,8 @@
+
+// send emails every second when in dev environment
+if (Meteor.absoluteUrl().indexOf('localhost') !== -1)
+  Herald.settings.queueTimer = 1000;
+  
 Meteor.startup(function () {
 
   Herald.collection.deny({
@@ -7,10 +12,6 @@ Meteor.startup(function () {
 
   // disable all email notifications when "emailNotifications" is set to false
   Herald.settings.overrides.email = !getSetting('emailNotifications', true);
-
-  // send emails every second when in dev environment
-  if (Meteor.absoluteUrl().indexOf('localhost') !== -1)
-    Herald.settings.queueTimer = 1000;
 
 });