+++ History.md
@@ -5,6 +5,8 @@
 * Added `threadModules` zone.
 * Added `upvoteCallbacks` and `downvoteCallbacks` callback arrays.
 * Fix “post awaiting moderation” message bug.
+* Added `postApproveCallbacks` callback array.
+* Added notifications for pending and approved posts, for admins and end users.
 
 ## v0.13.0 “ComponentScope”